Payments

+++ lapushki_api/Models/ClinicServices.cs
@@ -7,7 +7,7 @@ namespace lapushki_api.Models
     {
         [Key] public int id_service { get; set; }
         public string name { get; set; }
-        public double price { get; set; }
+        public decimal price { get; set; }
         public int duration { get; set; }
 
         [JsonIgnore]
